+++ sdk/storage/Azure.Storage.Files.Shares/tests/ServiceClientTests.cs
@@ -135,7 +135,7 @@ namespace Azure.Storage.Files.Shares.Tests
 
         [Test]
         [ServiceVersion(Min = ShareClientOptions.ServiceVersion.V2019_12_12)]
-        [PlaybackOnly("https://github.com/Azure/azure-sdk-for-net/issues/15505")]
+        [Ignore("https://github.com/Azure/azure-sdk-for-net/issues/15505")]
         [NonParallelizable]
         public async Task GetSetServicePropertiesAsync_SmbMultiChannel()
         {